create RoundResult component
